Change type

+++ 4thCourseS2/Parallel/ParcsJAVA/FuncMapMatrix/src/Map.java
@@ -2,10 +2,10 @@ import parcs.*;
 
 public class Map implements AM{
     public void run(AMInfo info){
-        long x,y,r;
-        x = info.parent.readLong();
+        int x,y,r;
+        x = info.parent.readInt();
         System.out.print(x);
-        y = info.parent.readLong();
+        y = info.parent.readInt();
         System.out.print('+');
         System.out.print(y);
         r = x+y;
